workflow等模块集成修复

Created-by: chenweizhao
Commit-by: chenweizhao_makeit
Merged-by: chenweizhao
Description: workflow等模块集成修复

See merge request: hw_makeit/agent-core-java!19

+++ src/main/java/com/openjiuwen/core/singleagent/agents/ReActAgentEvolve.java
@@ -243,8 +243,8 @@ public class ReActAgentEvolve extends BaseAgent {
         ContextWindow contextWindow = context.getContextWindow(
                 List.of(),
                 tools != null ? tools : null,
-                null,
-                null
+                (Integer) null,
+                (Integer) null
         );
 
         List<Object> skillMessages = new ArrayList<>(getSkillMessages());
